Moved code from backend/api/__init__.py to backend/api/core.py
